+++ lib/bindings/attribute.js
