+++ lib/bindings/child.js
