+++ lib/bindings/directive.js
