+++ lib/bindings/index.js
@@ -2,9 +2,9 @@ var walk = require('dom-walk');
 var each = require('each');
 var attrs = require('attributes');
 var domify = require('domify');
-var TextBinding = require('./text-binding');
-var AttrBinding = require('./attr-binding');
-var ChildBinding = require('./child-binding');
+var TextBinding = require('./text');
+var AttrBinding = require('./attribute');
+var ChildBinding = require('./child');
 var Directive = require('./directive');
 
 module.exports = function(options) {
