+++ lib/bindings/text.js
