+++ lib/index.js
@@ -3,7 +3,7 @@ var observer = require('path-observer');
 var type = require('type');
 var id = require('guid');
 var Interpolator = require('interpolate');
-var render = require('./render');
+var render = require('./bindings');
 
 /**
  * Run expressions
